feat: change to output tags instead of inline-styled elements
I decided on a separate release for this since it loses functionality
from the react-diff project (no longer displays with pink/green styles
when first installed).
The use-case here is a little different - a bland but stylable diff
that requires a little work to make it look ok, whereas the other
is more of a "just show me a diff with minimal effort" thing.